+++ Modules/GameEngineKit/Sources/_NewSystem/Exercises/Specialized/RemoteStandard/ReinforcerButton.swift
@@ -6,24 +6,33 @@ import DesignKit
 import RobotKit
 import SwiftUI
 
+// TODO(@ladislas): decide where to put this, keeping it here for now
+extension Robot.Reinforcer {
+
+    public func icon() -> Image {
+        switch self {
+            case .spinBlinkGreenOff:
+                return DesignKitAsset.Reinforcers.spinBlinkGreenOff.swiftUIImage
+            case .spinBlinkBlueViolet:
+                return DesignKitAsset.Reinforcers.spinBlinkBlueViolet.swiftUIImage
+            case .fire:
+                return DesignKitAsset.Reinforcers.fire.swiftUIImage
+            case .sprinkles:
+                return DesignKitAsset.Reinforcers.sprinkles.swiftUIImage
+            case .rainbow:
+                return DesignKitAsset.Reinforcers.rainbow.swiftUIImage
+        }
+    }
+
+}
+
 struct ReinforcerButton: View {
-    var reinforcer: Reinforcer
+    var reinforcer: Robot.Reinforcer
     let robot = Robot.shared
 
     var body: some View {
         Button {
-            switch reinforcer {
-                case .spinBlinkGreenOff:
-                    robot.run(.spinBlinkGreenOff)
-                case .spinBlinkBlueViolet:
-                    robot.run(.spinBlinkBlueViolet)
-                case .fire:
-                    robot.run(.fire)
-                case .sprinkles:
-                    robot.run(.sprinkles)
-                case .rainbow:
-                    robot.run(.rainbow)
-            }
+            robot.run(reinforcer)
         } label: {
             reinforcer.icon()
         }
